[maven-release-plugin] copy for tag fsl-1.3.1

+++ fsl-server/src/main/java/lsfusion/erp/utils/GetActiveJavaThreadsActionProperty.java
@@ -61,7 +61,7 @@ public class GetActiveJavaThreadsActionProperty extends ScriptingActionProperty 
 
         Map<Long, Thread> threadMap = ThreadUtils.getThreadMap();
         for (ThreadInfo threadInfo : threadsInfo) {
-            long id = threadInfo.getThreadId();
+            int id = (int) threadInfo.getThreadId();
             Thread thread = threadMap.get(id);
             DataObject currentObject = new DataObject(id);
